% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{check_data}
\usage{
check_data(
  data = c("batch1.csv", "batch2.csv"),
  sample.info = "sample.information.csv",
  path = "."
)
}
\arguments{
\item{data}{MS1 data tables.}

\item{sample.info}{Name of sample information.}

\item{path}{Work directory.}
}
\value{
Notice of data checking.
}
\description{
Check data format.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
