% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eic.R
\name{extract_eic}
\alias{extract_eic}
\title{extract_eic}
\usage{
extract_eic(
  targeted_table,
  object,
  polarity = c("positive", "negative"),
  mz_tolerance = 15,
  rt_tolerance = 30,
  threads = 5,
  add_point = FALSE,
  path = ".",
  group_for_figure = "QC",
  feature_type = c("pdf", "png"),
  width = 10,
  height = 6,
  device = NULL
)
}
\arguments{
\item{targeted_table}{A data.frame contains the variable_id,
mz and rt of features to extract.}

\item{object}{object from xcms. If you use massprocesser,
this should be the xdata3.}

\item{polarity}{positive or negative.}

\item{mz_tolerance}{default is 15 ppm.}

\item{rt_tolerance}{default is 30 ppm.}

\item{threads}{Number of threads.}

\item{add_point}{aad points in the EIC or not.}

\item{path}{Work directory.}

\item{group_for_figure}{what samples groups are used to extract EICs}

\item{feature_type}{png of pdf}

\item{width}{figure width}

\item{height}{height width}

\item{device}{Device to use.
Can either be a device function (e.g. png),
or one of "eps", "ps", "tex" (pictex), "pdf", "
jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
}
\value{
EICs.
}
\description{
Extract EICs of some features from some samples.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
