% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_massprocesser_logo.R
\name{get_massprocesser_version}
\alias{get_massprocesser_version}
\title{get_massprocesser_version}
\usage{
get_massprocesser_version()
}
\value{
massprocesser_version
}
\description{
get_massprocesser_version
}
\examples{
get_massprocesser_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
