% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_massprocesser_logo.R
\name{massprocesser_logo}
\alias{massprocesser_logo}
\title{massprocesser_logo}
\usage{
massprocesser_logo()
}
\value{
massprocesser_logo
}
\description{
massprocesser_logo
}
\examples{
massprocesser_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
