% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_adjusted_rt}
\alias{plot_adjusted_rt}
\title{plot_adjusted_rt}
\usage{
plot_adjusted_rt(object, title.size = 15, lab.size = 15, axis.text.size = 15)
}
\arguments{
\item{object}{object}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size}
}
\value{
A ggplot2 object.
}
\description{
plot_adjusted_rt
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
