% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_data}
\alias{process_data}
\title{process_data}
\usage{
process_data(
  path = ".",
  polarity = c("positive", "negative"),
  ppm = 15,
  peakwidth = c(5, 30),
  snthresh = 10,
  prefilter = c(3, 500),
  fitgauss = FALSE,
  integrate = 2,
  mzdiff = 0.01,
  noise = 500,
  threads = 6,
  binSize = 0.025,
  bw = 5,
  output_tic = TRUE,
  output_bpc = TRUE,
  output_rt_correction_plot = TRUE,
  min_fraction = 0.5,
  fill_peaks = FALSE,
  group_for_figure = "QC"
)
}
\arguments{
\item{path}{Work directory.}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{ppm}{see xcms.}

\item{peakwidth}{See xcms.}

\item{snthresh}{See xcms.}

\item{prefilter}{See xcms.}

\item{fitgauss}{see xcms.}

\item{integrate}{see xcms.}

\item{mzdiff}{see xcms.}

\item{noise}{See xcms.}

\item{threads}{Number of threads.}

\item{binSize}{see xcms.}

\item{bw}{see xcms.}

\item{output_tic}{Output TIC plot or not.}

\item{output_bpc}{Output BPC plot or not.}

\item{output_rt_correction_plot}{Output rt correction plot or not.}

\item{min_fraction}{See xcms.}

\item{fill_peaks}{Fill peaks NA or not.}

\item{group_for_figure}{Which group you want to use to output TIC and BPC and EIC. Default is QC.}
}
\value{
Peak table.
}
\description{
Raw MS data processing using xcms.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
