% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_targeted_table}
\alias{check_targeted_table}
\title{check_targeted_table}
\usage{
check_targeted_table(targeted_table)
}
\arguments{
\item{targeted_table}{targeted_table, 3 columns, 1. variable_id, 2. mz, 3. rt..}
}
\value{
Notice of data checking.
}
\description{
Check targeted_table for extract_eic function.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
