% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_adjusted_rt}
\alias{plot_adjusted_rt}
\title{plot_adjusted_rt}
\usage{
plot_adjusted_rt(
  object,
  title.size = 13,
  lab.size = 13,
  axis.text.size = 12,
  interactive = FALSE,
  group_for_figure = "QC",
  sample_for_figure = NULL
)
}
\arguments{
\item{object}{object}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size}

\item{interactive}{interactive or not.}

\item{group_for_figure}{What groups to show EIC.}

\item{sample_for_figure}{sample_for_figure}
}
\value{
A ggplot2 object.
}
\description{
plot_adjusted_rt
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
